pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Ganesh Rao <mister.xrao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.13"
fastapi = {extras = ["standard"], version = "^0.115.6"}
langchain-google-genai = "^2.0.7"
langchain-core = "^0.3.28"
langchain-community = "^0.3.13"
langgraph = "^0.2.60"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
